backend/requirements.txt
flask==3.1.1
flask-cors==6.0.1
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.8.0
requests==2.32.3
sqlalchemy==2.0.41
python-dotenv==1.2.1
PyJWT==2.9.0
gunicorn==23.0.0
